fix(perf): extract recordSpawnTiming into its own module so it can be tested
The recordSpawnTiming util and its _perfEnabled cache used to live in tabStore.ts. Its test (src/test/recordSpawnTiming.test.ts) tried to exercise both perf=true and perf=false in the same file by calling vi.resetModules() + dynamic import('../stores/tabStore') per describe block. Under vitest 3.x + jsdom, that dynamic import hung indefinitely after vi.resetModules — tabStore pulls the entire Terminal barrel (xterm, monaco) through a circular dependency (tabStore -> Terminal -> useTerminal -> tabStore) that does not survive module-cache reset cleanly. Every test in the file timed out at 5000ms.

Fix: move recordSpawnTiming + the perf cache into src/utils/recordSpawnTiming.ts (no transitive dependencies beyond @tauri-apps/api). tabStore imports the new module. The test now imports recordSpawnTiming directly and calls it without going through addTab() — vi.resetModules() only has to reset the perf cache, not the entire UI tree.

tabStore.test.ts (42 tests) still passes — public behavior of addTab/duplicateTab/splitPane unchanged. recordSpawnTiming.test.ts (3 tests) now passes end-to-end.

[CLEAN-ARCH] separation of concerns - perf instrumentation is no longer entangled with tab state management
