+++ src/test/recordSpawnTiming.test.ts
@@ -6,34 +6,49 @@
  * - Listens for pty-perf event when perf is enabled
  * - One-shot: unlisten after matching session event
  *
- * NOTE: checkPerfEnabled() caches the result at module level, so only
- * the FIRST call per test file determines the perf state. Tests that
- * need perf=true run first, then perf=false tests verify via separate
- * assertions. Each describe block uses vi.resetModules() for isolation.
+ * checkPerfEnabled() caches its result at module level. To exercise
+ * both perf=on and perf=off in the same file we call `vi.resetModules()`
+ * before each `it` and dynamically re-import the module under test —
+ * which is `../utils/recordSpawnTiming`, NOT `../stores/tabStore`. The
+ * old test imported tabStore through `vi.doMock` + a dynamic import;
+ * because tabStore pulls in the entire Terminal barrel (xterm, monaco,
+ * a tabStore↔useTerminal cycle), that dynamic import hung after
+ * `vi.resetModules()`. Extracting recordSpawnTiming into its own module
+ * (commit accompanying this test rewrite) lets us reset only the timer
+ * cache and test the function directly.
  *
  * Tags: [TDD], [Fix-8]
  */
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { act } from "@testing-library/react";
 
-// Shared mock references — reset per describe block via resetModules
-let mockInvoke: ReturnType<typeof vi.fn>;
-let mockListen: ReturnType<typeof vi.fn>;
+// Hoisted mock fns — `vi.mock` factories run before any imports, so
+// the captured refs must live in `vi.hoisted`.
+const { mockInvoke, mockListen } = vi.hoisted(() => ({
+  mockInvoke: vi.fn(),
+  mockListen: vi.fn().mockResolvedValue(vi.fn()),
+}));
+
+vi.mock("@tauri-apps/api/core", () => ({
+  invoke: (...args: unknown[]) => mockInvoke(...args),
+}));
+
+vi.mock("@tauri-apps/api/event", () => ({
+  listen: (...args: unknown[]) => mockListen(...args),
+}));
+
+/** Drains microtasks + macrotask queue so the fire-and-forget
+ *  recordSpawnTiming async block can observably reach its IPC calls
+ *  before we assert. Using setTimeout(0) yields to the macrotask queue. */
+async function flushAsync(): Promise<void> {
+  await new Promise((resolve) => setTimeout(resolve, 0));
+  await new Promise((resolve) => setTimeout(resolve, 0));
+}
 
 describe("recordSpawnTiming (perf disabled)", () => {
   beforeEach(async () => {
     vi.resetModules();
-
-    mockInvoke = vi.fn();
-    mockListen = vi.fn().mockResolvedValue(vi.fn());
-
-    vi.doMock("@tauri-apps/api/core", () => ({
-      invoke: (...args: unknown[]) => mockInvoke(...args),
-    }));
-
-    vi.doMock("@tauri-apps/api/event", () => ({
-      listen: (...args: unknown[]) => mockListen(...args),
-    }));
+    mockInvoke.mockReset();
+    mockListen.mockReset().mockResolvedValue(vi.fn());
   });
 
   afterEach(() => {
@@ -41,24 +56,19 @@ describe("recordSpawnTiming (perf disabled)", () => {
   });
 
   it("does not register listener when perf_enabled returns false", async () => {
-    // perf_enabled returns false
     mockInvoke.mockImplementation((cmd: string) => {
-      if (cmd === "pty_spawn") return Promise.resolve("session-perf-off");
       if (cmd === "perf_enabled") return Promise.resolve(false);
       return Promise.resolve(undefined);
     });
 
-    const { useTabStore } = await import("../stores/tabStore");
-    useTabStore.setState({ tabs: [], activeTabId: "", tabCounter: 0 });
+    const { recordSpawnTiming } = await import("../utils/recordSpawnTiming");
+    recordSpawnTiming("session-perf-off", 0);
 
-    await act(async () => {
-      await useTabStore.getState().addTab();
-    });
-
-    // Flush the async void recordSpawnTiming
+    // Flush the fire-and-forget async block.
     await vi.waitFor(() => {
       expect(mockInvoke).toHaveBeenCalledWith("perf_enabled");
     });
+    await flushAsync();
 
     // listen should NOT have been called for pty-perf
     const perfListenCalls = mockListen.mock.calls.filter(
@@ -71,17 +81,8 @@ describe("recordSpawnTiming (perf disabled)", () => {
 describe("recordSpawnTiming (perf enabled)", () => {
   beforeEach(async () => {
     vi.resetModules();
-
-    mockInvoke = vi.fn();
-    mockListen = vi.fn().mockResolvedValue(vi.fn());
-
-    vi.doMock("@tauri-apps/api/core", () => ({
-      invoke: (...args: unknown[]) => mockInvoke(...args),
-    }));
-
-    vi.doMock("@tauri-apps/api/event", () => ({
-      listen: (...args: unknown[]) => mockListen(...args),
-    }));
+    mockInvoke.mockReset();
+    mockListen.mockReset().mockResolvedValue(vi.fn());
   });
 
   afterEach(() => {
@@ -90,19 +91,13 @@ describe("recordSpawnTiming (perf enabled)", () => {
 
   it("registers pty-perf listener when perf_enabled returns true", async () => {
     mockInvoke.mockImplementation((cmd: string) => {
-      if (cmd === "pty_spawn") return Promise.resolve("session-perf-on");
       if (cmd === "perf_enabled") return Promise.resolve(true);
       return Promise.resolve(undefined);
     });
 
-    const { useTabStore } = await import("../stores/tabStore");
-    useTabStore.setState({ tabs: [], activeTabId: "", tabCounter: 0 });
+    const { recordSpawnTiming } = await import("../utils/recordSpawnTiming");
+    recordSpawnTiming("session-perf-on", 0);
 
-    await act(async () => {
-      await useTabStore.getState().addTab();
-    });
-
-    // Flush the async void recordSpawnTiming
     await vi.waitFor(() => {
       const perfListenCalls = mockListen.mock.calls.filter(
         (call: unknown[]) => call[0] === "pty-perf",
@@ -116,17 +111,12 @@ describe("recordSpawnTiming (perf enabled)", () => {
     mockListen.mockResolvedValue(mockUnlisten);
 
     mockInvoke.mockImplementation((cmd: string) => {
-      if (cmd === "pty_spawn") return Promise.resolve("session-abc12345");
       if (cmd === "perf_enabled") return Promise.resolve(true);
       return Promise.resolve(undefined);
     });
 
-    const { useTabStore } = await import("../stores/tabStore");
-    useTabStore.setState({ tabs: [], activeTabId: "", tabCounter: 0 });
-
-    await act(async () => {
-      await useTabStore.getState().addTab();
-    });
+    const { recordSpawnTiming } = await import("../utils/recordSpawnTiming");
+    recordSpawnTiming("session-abc12345", 0);
 
     // Wait for pty-perf listener to be registered
     await vi.waitFor(() => {
@@ -136,7 +126,7 @@ describe("recordSpawnTiming (perf enabled)", () => {
       expect(perfListenCalls).toHaveLength(1);
     });
 
-    // Get the callback that was registered
+    // Grab the callback that was registered
     const perfListenCall = mockListen.mock.calls.find(
       (call: unknown[]) => call[0] === "pty-perf",
     );
